# api.yaml
openapi: 3.0.0
info:
  description: Документация к нашему REST чудовищу
  version: 1.0.0
  title: Drello backend API
...
servers:
  - url: http://localhost:8080
security:
  - app_id: []
...
paths:
...
  /session:
    post:
      tags:
        - session
      summary: Операция авторизации в нашем сервисе
      description: ...
      requestBody:
        description: Данные, необходимые для авторизации
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Board/properties/admins/items"
            example:
              nickname: Legolas1
              password: qwerty1234
      responses:
        "200":
          description: Успешно авторизовались, в ответе вшита кука
          headers:
            session_id:
              schema:
                type: string
              description: Наша авторизационная кука.
        "303":
          description: Постучались на данный endpoint с выставленной кукой
        "400":
          description: Невалидное тело запроса
        "412":
          description: Неверный пароль
    get:
      tags:
        - session
      summary: Проверяем наличие сессии
      description: ...
      parameters:
        - in: cookie
          name: session_id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: В запросе выставлена актуальная кука, всё ок
        "401":
          description: В запросе куки нет
        "500":
          description: На сервере ошибка cоответствия куки пользователя и кукой, которая
            лежит в базе (пока в работе)
    delete:
      tags:
        - session
      summary: Операция logout, удаляем куку
      description: ...
      parameters:
        - in: cookie
          name: session_id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Успешно удалили куку
        "401":
          description: В запросе куки нет
        "500":
          description: На сервере ошибка удаления куки из хранилища
...
  "/boards/{bid}/columns":
    get:
      tags:
        - columns
      summary: Получаем все колонки конкретной доски
      parameters:
        - in: cookie
          name: session_id
          schema:
            type: string
          required: true
        - in: path
          name: bid
          required: true
          schema:
            type: integer
          description: Уникальный номер доски
          example: 10
      responses:
        "200":
          description: Успешно получили колонки
          content:
            application/json:
              schema:
                type: array
                items:
                  - $ref: "#/components/schemas/Board/properties/columns/items"
              example:
                columns:
                  - id: 152
                    title: Homework
                    pos: "0.08266362238478808"
        "400":
          description: Неверные параметры запроса
        "401":
          description: В запросе отсутствует кука
        "403":
          description: Недостаточно прав для совершения операции
        "500":
          description: Внутренняя ошибка сервера
    post:
      tags:
        - columns
      summary: Создаем колонку на выбранной доске
      parameters:
        - in: cookie
          name: session_id
          schema:
            type: string
          required: true
        - in: path
          name: bid
          required: true
          schema:
            type: integer
          description: Уникальный номер доски
          example: 10
      requestBody:
        description: Данные, необходимые для создания колонки
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Board/properties/columns/items"
            example:
              title: TitleForColumn
              pos: "0.08266362238478808"
      responses:
        "200":
          description: Успешно создали колонку
          content:
            application/json:
              schema:
                - $ref: "#/components/schemas/Column/properties/tasks/items"
              example:
                id: 110
                title: TitleForColumn
                pos: "0.08266362238478808"
        "400":
          description: Неверные параметры запроса
        "401":
          description: В запросе отсутствует кука
        "403":
          description: Недостаточно прав для совершения операции
        "500":
          description: Внутренняя ошибка сервера
  "/boards/{bid}/columns/{cid}":
    get:
      tags:
        - columns
      summary: Получаем конкретную колонку конкретной доски
      parameters:
        - in: cookie
          name: session_id
          schema:
            type: string
          required: true
        - in: path
          name: bid
          required: true
          schema:
            type: integer
          description: Уникальный номер доски
          example: 10
        - in: path
          name: сid
          required: true
          schema:
            type: integer
          description: Уникальный номер колонки
          example: 3
      responses:
        "200":
          description: Успешно получили колонку
          content:
            application/json:
              schema:
                - $ref: "#/components/schemas/Board/properties/columns/items"
              example:
                id: 5
                title: Review
                pos: 0.08266362238478808
        "400":
          description: Неверные параметры запроса
        "401":
          description: В запросе отсутствует кука
        "403":
          description: Недостаточно прав для совершения операции
        "500":
          description: Внутренняя ошибка сервера
    put:
      tags:
        - columns
      summary: Обновляем конкретную колонку конкретной доски (ЕЩЁ В ПРОЦЕССЕ РЕАЛИЗАЦИИ)
      parameters:
        - in: cookie
          name: session_id
          schema:
            type: string
          required: true
        - in: path
          name: bid
          required: true
          schema:
            type: integer
          description: Уникальный номер доски
          example: 10
        - in: path
          name: сid
          required: true
          schema:
            type: integer
          description: Уникальный номер колонки
          example: 3
      requestBody:
        description: Данные, необходимые для обновления колонки
        required: true
        content:
          application/json:
            schema:
              - $ref: "#/components/schemas/Board/properties/columns/items"
            example:
              title: NewTitle
              pos: 0.08266362238478808
      responses:
        "200":
          description: Успешно обновили колонку
        "400":
          description: Неверные параметры запроса
        "401":
          description: В запросе отсутствует кука
        "403":
          description: Недостаточно прав для совершения операции
        "500":
          description: Внутренняя ошибка сервера
    delete:
      tags:
        - columns
      summary: Удаляем конкретную колонку конкретной доски
      parameters:
        - in: cookie
          name: session_id
          schema:
            type: string
          required: true
        - in: path
          name: bid
          required: true
          schema:
            type: integer
          description: Уникальный номер доски
          example: 10
        - in: path
          name: сid
          required: true
          schema:
            type: integer
          description: Уникальный номер колонки
          example: 3
      responses:
        "200":
          description: Успешно удалили доску
        "400":
          description: Неверные параметры запроса
        "401":
          description: В запросе отсутствует кука
        "403":
          description: Недостаточно прав для совершения операции
        "500":
          description: Внутренняя ошибка сервера
  "/boards/{bid}/columns/{cid}/tasks":
    get:
      tags:
        - tasks
      summary: Получаем таски определенных доски и колонки
      parameters:
        - in: cookie
          name: session_id
          schema:
            type: string
          required: true
        - in: path
          name: bid
          required: true
          schema:
            type: integer
          description: Уникальный номер доски
          example: 10
        - in: path
          name: cid
          required: true
          schema:
            type: integer
          description: Уникальный номер колонки
          example: 3
      responses:
        "200":
          description: Успешно получили таски
          content:
            application/json:
              schema:
                type: array
                items:
                  - $ref: "#/components/schemas/Column/properties/tasks/items"
              example:
                tasks:
                  - id: 152
                    title: Homework
                    description: HW for doing
                    level: 10
                    deadline: 2017-07-21
                    pos: "0.08266362238478808"
                    cid: 11
        "400":
          description: Неверные параметры запроса
        "401":
          description: В запросе отсутствует кука
        "403":
          description: Недостаточно прав для совершения операции
        "500":
          description: Внутренняя ошибка сервера
    post:
      tags:
        - tasks
      summary: Создаем таску на определенной доске и колонке
      parameters:
        - in: cookie
          name: session_id
          schema:
            type: string
          required: true
        - in: path
          name: bid
          required: true
          schema:
            type: integer
          description: Уникальный номер доски
          example: 10
        - in: path
          name: cid
          required: true
          schema:
            type: integer
          description: Уникальный номер колонки
          example: 3
      requestBody:
        description: Данные, необходимые для создания таски
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Column/properties/tasks/items"
            example:
              title: Homework
              description: HW for doing
              level: 10
              deadline: "20170721"
              pos: 0.08266362238478808
      responses:
        "200":
          description: Успешно создали таску
          content:
            application/json:
              schema:
                - $ref: "#/components/schemas/Column/properties/tasks/items"
              example:
                id: 25
                title: Homework
                description: HW for doing
                level: 10
                deadline: "20170721"
                pos: 0.08266362238478808
        "400":
          description: Неверные параметры запроса
        "401":
          description: В запросе отсутствует кука
        "403":
          description: Недостаточно прав для совершения операции
        "500":
          description: Внутренняя ошибка сервера
  "/boards/{bid}/columns/{cid}/tasks/{tid}":
    get:
      tags:
        - tasks
      summary: Получаем определенную таску определенных доски и колонки
      parameters:
        - in: cookie
          name: session_id
          schema:
            type: string
          required: true
        - in: path
          name: bid
          required: true
          schema:
            type: integer
          description: Уникальный номер доски
          example: 10
        - in: path
          name: cid
          required: true
          schema:
            type: integer
          description: Уникальный номер колонки
          example: 3
        - in: path
          name: tid
          required: true
          schema:
            type: integer
          description: Уникальный номер таски
          example: 8
      responses:
        "200":
          description: Успешно получили таску
          content:
            application/json:
              schema:
                - $ref: "#/components/schemas/Column/properties/tasks/items"
              example:
                task:
                  id: 152
                  title: Homework
                  description: HW for doing
                  level: 10
                  deadline: 2017-07-21
                  pos: 0.08266362238478808
                  cid: 11
        "400":
          description: Неверные параметры запроса
        "401":
          description: В запросе отсутствует кука
        "403":
          description: Недостаточно прав для совершения операции
        "500":
          description: Внутренняя ошибка сервера
    put:
      tags:
        - tasks
      summary: Обновляем определенную таску определенных доски и колонки
      parameters:
        - in: cookie
          name: session_id
          schema:
            type: string
          required: true
        - in: path
          name: bid
          required: true
          schema:
            type: integer
          description: Уникальный номер доски
          example: 10
        - in: path
          name: cid
          required: true
          schema:
            type: integer
          description: Уникальный номер колонки
          example: 3
        - in: path
          name: tid
          required: true
          schema:
            type: integer
          description: Уникальный номер таски
          example: 8
      requestBody:
        description: Данные, необходимые для обновления таски
        required: true
        content:
          application/json:
            schema:
              - $ref: "#/components/schemas/Column/properties/tasks/items"
            example:
              title: Homework
              description: HW for doing
              level: 10
              deadline: 2017-07-21
              pos: 0.082663622384788
              cid: 11
      responses:
        "200":
          description: Успешно обновили таску
        "400":
          description: Неверные параметры запроса
        "401":
          description: В запросе отсутствует кука
        "403":
          description: Недостаточно прав для совершения операции
        "500":
          description: Внутренняя ошибка сервера
    delete:
      tags:
        - tasks
      summary: Удаляем определенную таску определенных доски и колонки
      parameters:
        - in: cookie
          name: session_id
          schema:
            type: string
          required: true
        - in: path
          name: bid
          required: true
          schema:
            type: integer
          description: Уникальный номер доски
          example: 10
        - in: path
          name: cid
          required: true
          schema:
            type: integer
          description: Уникальный номер колонки
          example: 3
        - in: path
          name: tid
          required: true
          schema:
            type: integer
          description: Уникальный номер таски
          example: 8
      responses:
        "200":
          description: Успешно удалили таску
        "400":
          description: Неверные параметры запроса
        "401":
          description: В запросе отсутствует кука
        "403":
          description: Недостаточно прав для совершения операции
        "500":
          description: Внутренняя ошибка сервера
components:
...
  securitySchemes:
    app_id:
      type: apiKey
      description: API key to authorize requests. If you don't have an OpenWeatherMap
        API key, use `fd4698c940c6d1da602a70ac34f0b147`.
      name: appid
      in: query